+++ src/Controller/Component/FlashComponent.php
@@ -41,8 +41,13 @@ class FlashComponent extends CakeFlashComponent {
 	 * @uses MeTools\Core\Plugin::path()
 	 */
 	public function __call($name, $args) {
-		if(!isset($args[1]['plugin']) && is_readable(Plugin::path('MeTools', 'src'.DS.'Template'.DS.'Element'.DS.'Flash'.DS.Inflector::underscore($name).'.ctp')))
-			$args[1]['plugin'] = 'MeTools';
+        if(!isset($args[1]['plugin'])) {
+            $file = Plugin::path('MeTools', 'src'.DS.'Template'.DS.'Element'.DS.'Flash'.DS.Inflector::underscore($name).'.ctp', TRUE);
+            
+            if($file) {
+                $args[1]['plugin'] = 'MeTools';
+            }
+        }
 		
 		return parent::__call($name, $args);
 	}