+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 2.x branch
 ## 2.6 branch
 ### 2.6.0
-* removed `Xml` class;
-* removed `SecurityComponent` class.
+* improved the `Plugin` class;
+* removed the `Xml` class;
+* removed the `SecurityComponent` class.
 
 ## 2.5 branch
 ### 2.5.2